+++ webpack.config.js
@@ -1,37 +1,37 @@
 const path = require('path');
 
 module.exports = (env, argv) => {
-  const isDev = argv.mode === 'development';
-  
+  const isDev = argv && argv.mode === 'development';
+
   return {
     entry: {
       'gridstack-all': './src/gridstack.ts',
     },
-    mode: isDev ? 'development' : 'production',
-    devtool: isDev ? 'eval-source-map' : 'source-map',
+    mode: isDev ? 'development' : 'production', // production vs development
+    devtool: isDev ? 'eval-source-map' : 'source-map', // for best (large .js) debugging
     module: {
-    rules: [
-      {
-        test: /\.ts$/,
-        use: {
-          loader: 'ts-loader',
-          options: {
-            configFile: 'tsconfig.build.json'
-          }
+      rules: [
+        {
+          test: /\.ts$/,
+          use: {
+            loader: 'ts-loader',
+            options: {
+              configFile: 'tsconfig.build.json'
+            }
+          },
+          exclude: ['/node_modules/', '/spec/'],
         },
-        exclude: ['/node_modules/', '/spec/'],
-      },
-    ],
-  },
-  resolve: {
-    extensions: [ '.ts', '.js' ],
-  },
-  output: {
-    filename: '[name].js',
-    path: path.resolve(__dirname, 'dist'),
-    library: 'GridStack',
-    libraryExport: 'GridStack',
-    libraryTarget: 'umd', // var|assign|this|window|self|global|commonjs|commonjs2|commonjs-module|amd|amd-require|umd|umd2|jsonp|system
-  }
+      ],
+    },
+    resolve: {
+      extensions: [ '.ts', '.js' ],
+    },
+    output: {
+      filename: '[name].js',
+      path: path.resolve(__dirname, 'dist'),
+      library: 'GridStack',
+      libraryExport: 'GridStack',
+      libraryTarget: 'umd', // var|assign|this|window|self|global|commonjs|commonjs2|commonjs-module|amd|amd-require|umd|umd2|jsonp|system
+    }
   };
 };